SynchroPhasor: openPDCManager - Bug fixes for OutputStream wizard and other pages against SQL script and blank database.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/openpdc/Synchrophasor/Current Version;C35008

+++ Source/Applications/openPDCManager/Silverlight/ModalDialogs/OutputStreamWizard/CurrentDevices.xaml.cs
@@ -272,7 +272,10 @@ namespace openPDCManager.Silverlight.ModalDialogs.OutputStreamWizard
 		}
 		void ButtonDelete_Click(object sender, RoutedEventArgs e)
 		{
-			client.DeleteOutputStreamDeviceAsync(sourceOutputStreamID, devicesToBeDeleted);
+			if (devicesToBeDeleted.Count > 0)
+				client.DeleteOutputStreamDeviceAsync(sourceOutputStreamID, devicesToBeDeleted);
+			else
+				MessageBox.Show("Select atleast one device.");
 		}
 		void ButtonAdd_Click(object sender, RoutedEventArgs e)
 		{
